+++ src/Exception/InvalidConfigException.php
@@ -1,11 +1,12 @@
 <?php
+
 /**
- * @see       https://github.com/zendframework/zend-expressive-authorization-acl for the canonical source repository
- * @copyright Copyright (c) 2017 Zend Technologies USA Inc. (http://www.zend.com)
- * @license   https://github.com/zendframework/zend-expressive-authorization-acl/blob/master/LICENSE.md New BSD License
+ * @see       https://github.com/mezzio/mezzio-authorization-acl for the canonical source repository
+ * @copyright https://github.com/mezzio/mezzio-authorization-acl/blob/master/COPYRIGHT.md
+ * @license   https://github.com/mezzio/mezzio-authorization-acl/blob/master/LICENSE.md New BSD License
  */
 
-namespace Zend\Expressive\Authorization\Acl\Exception;
+namespace Mezzio\Authorization\Acl\Exception;
 
 class InvalidConfigException extends RuntimeException implements ExceptionInterface
 {